Refactor function names and argument type in sqlbuilder_test
This commit does two main things. First, it changes three function names from capitalized to lowercase, namely `EscapeStr` to `escapeStr` and `RowBuild` to `rowBuild`. This is to adhere to Go's convention of unexported identifiers. These functions do not need to be accessed from other packages so they should not be exported.

Second, it modifies the argument passed to `QueriesBuild` function from `tt.args.maxallowpack` to `uint64(tt.args.maxallowpack)`. The reason behind this is to match the expected data type in `QueriesBuild` function's parameters and thereby preventing any potential type mismatch errors.