Remove max packet size constraints in sqlbuilder
The constraints linked to the min/max size of the allowed packets in the sqlbuilder script have been removed. These limitations were superfluous in the scope of the sqlbuilder's functionality and caused unnecessary errors. Instead, the responsibility to manage packet sizes should be at a higher level of the application architecture.