Update error handling for large SQL queries
Updated the SQL builder to split large SQL queries into several smaller ones when they exceed the max_allowed_packet limit. This is done by appending to the 'queries' array and resetting the 'outValuesString' buffer. An additional modification was made to improve the format of the error message which will now also include the current size of the query when a limit is reached. This will provide more clarity when debugging large queries.